Move all build-related code to build directory
